Added solidity file output and removed Makefile dependency

+++ examples/simple/ConstAndImmutable.sol
@@ -1,18 +1,16 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity >=0.7.4;
-
-uint constant X = 32**22 + 8;
+uint256 constant X = 32**22 + 8;
 
 contract ConstAndImmutable {
     string constant TEXT = "abc";
     bytes32 constant MY_HASH = keccak256("abc");
-    uint immutable decimals;
-    uint immutable maxBalance;
+    uint256 immutable decimals;
+    uint256 immutable maxBalance;
     address immutable owner = msg.sender;
 
-    constructor(uint _decimals, address _reference) {
+    constructor(uint256 _decimals, address _reference) {
         decimals = _decimals;
-        // Assignments to immutables can even access the environment.
         maxBalance = _reference.balance;
     }
