+++ examples/simple/TryCatch.sol
@@ -1,32 +1,25 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity >0.8.0;
 
-interface DataFeed { function getData(address token) external returns (uint value); }
+interface DataFeed {
+    function getData(address token) external returns (uint256 value);
+}
 
 contract FeedConsumer {
     DataFeed feed;
-    uint errorCount;
-    function rate(address token) public returns (uint value, bool success) {
-        // Permanently disable the mechanism if there are
-        // more than 10 errors.
+    uint256 errorCount;
+
+    function rate(address token) public returns (uint256 value, bool success) {
         require(errorCount < 10);
-        try feed.getData(token) returns (uint v) {
+        try feed.getData(token) returns (uint256 v) {
             return (v, true);
-        } catch Error(string memory /*reason*/) {
-            // This is executed in case
-            // revert was called inside getData
-            // and a reason string was provided.
+        } catch Error(string memory) {
             errorCount++;
             return (0, false);
-        } catch Panic(uint /*errorCode*/) {
-            // This is executed in case of a panic,
-            // i.e. a serious error like division by zero
-            // or overflow. The error code can be used
-            // to determine the kind of error.
+        } catch Panic(uint256) {
             errorCount++;
             return (0, false);
-        } catch (bytes memory /*lowLevelData*/) {
-            // This is executed in case revert() was used.
+        } catch (bytes memory) {
             errorCount++;
             return (0, false);
         }
