+++ examples/simple/ViewAndPure.sol
@@ -2,11 +2,11 @@
 pragma solidity >=0.5.0 <0.9.0;
 
 contract C {
-    function f(uint a, uint b) public view returns (uint) {
+    function f(uint256 a, uint256 b) public view returns (uint256) {
         return a * (b + 42) + block.timestamp;
     }
 
-    function g(uint a, uint b) public pure returns (uint) {
+    function g(uint256 a, uint256 b) public pure returns (uint256) {
         return a * (b + 42);
     }
 }
